ap calculations
added functions for ap calculations
added a "starting s. lv" property to defaultcharacterstate
added an ap and s. lv properties to characterstate
the kill event has a different syntax and adds ap to the credited characters
added a parse_character_ap parsing function
fixes in the notes to reflect the changes